changing servers and removing client
