+++ src/polhemus_tf_broadcaster.cpp
@@ -32,14 +32,15 @@
 #include <ctype.h>
 #include <string.h>
 #include <signal.h>
+#include <mutex>
 
 #include <sys/time.h>
 #include <time.h>
 #include <ros/ros.h>
 #include <tf2_ros/transform_broadcaster.h>
 #include <geometry_msgs/TransformStamped.h>
-#include "polhemus_ros_driver/calibrate.h"
 
+#include "polhemus_ros_driver/calibrate.h"
 #include "liberty.h"
 #include "protocol.h"
 
@@ -64,6 +65,12 @@ static int count_bits(uint16_t v) {
 
 /* main loop running? */
 static int go_on;
+
+/* mutex for liberty_send */
+std::mutex mtx;
+
+/* Handle to communicate with device*/
+usb_dev_handle *handle;
 
 static void signal_handler(int s) {
   switch (s) {
@@ -126,12 +133,33 @@ static void set_hemisphere(usb_dev_handle *handle, int x, int y, int z) {
   liberty_send(handle, cmd);
 }
 
+/* Calibrate the sensors: causes the sensor to be electronically aligned in orientation (Azimuth Reference,
+Elevation Reference, Roll Reference) with the user system coordinates, and establishes the boresight
+reference angles for the station. */
+bool calibrate(polhemus_ros_driver::calibrate::Request  &req, polhemus_ros_driver::calibrate::Response &res)
+{
+  if (!handle) {
+    fprintf(stderr, "Could not get a handle to the Polhemus Liberty device.\n");
+    res.success = false;
+    return false;
+  }
+  std::string cmd = "";
+  cmd = "b" + req.station + "," + req.azref + "," + req.elref + "," + req.rlref +  "," ;
+  cmd += req.reset_origin ? "1" : "0";
+  cmd += "\r";
+  ROS_INFO("Calibration request: %s\n",cmd.c_str());
+  mtx.lock();
+  liberty_send(handle, (char *)cmd.c_str());
+  mtx.unlock();
+  res.success = true;
+  return true;
+}
+
 int main(int argc, char** argv) {
 
   int i, nstations;
   double x_hs, y_hs, z_hs;
   struct usb_device *dev;
-  usb_dev_handle *handle;
   buffer_t buf;
   struct timeval tv;
 
@@ -170,28 +198,9 @@ int main(int argc, char** argv) {
   ros::init(argc, argv, "polhemus_tf_broadcaster");
   ros::NodeHandle nh;
 
-  /* Calibrate the sensors: causes the sensor to be electronically aligned in orientation (Azimuth Reference,
-  Elevation Reference, Roll Reference) with the user system coordinates, and establishes the boresight
-  reference angles for the station. */
-  ros::ServiceClient client = nh.serviceClient<polhemus_ros_driver::calibrate>("calibration");
-  polhemus_ros_driver::calibrate calibration_srv;
-  calibration_srv.request.station = "*";
-  calibration_srv.request.azref = "";
-  calibration_srv.request.elref = "";
-  calibration_srv.request.rlref = "0";
-  calibration_srv.request.reset_origin = false;
-  if (client.call(calibration_srv))
-  {
-    if (calibration_srv.response.success == 0)
-    {
-      ROS_INFO("Sensors were calibrated successfully: %d", calibration_srv.response.success);
-    }
-  }
-  else
-  {
-    ROS_ERROR("Failed to call calibration service");
-    return 1;
-  }
+  // Calibration service
+  ros::ServiceServer service = nh.advertiseService("calibration", calibrate);
+  ROS_INFO("Service ready to calibrate the sensors.");
 
 
   /* define which information to get per sensor (called a station
